requirements.txt
# MapBench.Live Python Dependencies

# Core dependencies
pyyaml>=6.0
python-dotenv>=1.0.0
aiofiles>=23.0.0

# API clients
openai>=1.0.0
google-cloud-aiplatform>=1.38.0
anthropic>=0.18.0

# Web framework (for local API if needed)
fastapi>=0.109.0
uvicorn>=0.25.0

# Data processing
pandas>=2.0.0
pillow>=10.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0